SearchBox: refactor - add more types for geocoder, move selected factories
