+++ src/components/SearchBox/types.ts
@@ -1,40 +1,46 @@
 import { LonLat } from '../../services/types';
 import { Star } from '../utils/StarsContext';
 import { ASTNode } from './queryWizard/ast';
+import { BBox } from 'geojson';
+import { Bbox } from '../utils/MapStateContext';
 
 type GenericOption<T extends string, U extends Object | null> = {
   type: T;
 } & (U extends null ? {} : Record<T, U>);
 
-export type GeocoderOption = GenericOption<
-  'geocoder',
-  {
-    loading: true;
-    skeleton: true;
-    nonOsmObject: true;
-    osmMeta: { type: string; id: number };
-    center: LonLat;
-    tags: Record<string, string>;
-    properties?: {
-      class: string | null;
-      place: string | null | undefined;
-      street: string | null | undefined;
-      city: string | null | undefined;
-      housenumber: string | null | undefined;
-      streetnumber: string | null | undefined;
-      osm_key: string;
-      osm_value: string;
-      osm_type: string;
-      osm_id: string;
-      name: string;
-      /** Either a string or an array */
-      extent: any;
-    };
-    geometry: {
-      coordinates: LonLat;
-    };
-  }
->;
+export type PhotonGeojsonFeature = {
+  type: 'Feature';
+  properties: {
+    type: string; // house, street, city
+    place: string | null | undefined;
+    street: string | null | undefined;
+    housenumber: string | null | undefined;
+    streetnumber: string | null | undefined;
+    city: string | null | undefined;
+    district?: string;
+    county?: string; // not sure if it ever appears
+    state?: string;
+    locality?: string;
+    postcode?: string;
+    country?: string;
+    countrycode?: string;
+    osm_key: string;
+    osm_value: string;
+    osm_type: string;
+    osm_id: string;
+    name: string;
+    extent?: Bbox;
+  };
+  geometry: {
+    coordinates: LonLat;
+  };
+};
+
+export type PhotonResponse = {
+  features: PhotonGeojsonFeature[];
+};
+
+export type GeocoderOption = GenericOption<'geocoder', PhotonGeojsonFeature>;
 
 export type OverpassOption = GenericOption<
   'overpass',
